[td] Use shared filebrowser in notebook
