+++ mage_ai/frontend/components/PipelineDetail/FileHeaderMenu/index.tsx
@@ -1,5 +1,6 @@
 import { useEffect, useMemo, useRef, useState } from 'react';
 
+import Button from '@oracle/elements/Button';
 import ClickOutside from '@oracle/components/ClickOutside';
 import FlexContainer from '@oracle/components/FlexContainer';
 import FlyoutMenu from '@oracle/components/FlyoutMenu';
@@ -365,7 +366,12 @@ function FileHeaderMenu({
     >
       <FlexContainer>
         <div style={{ position: 'relative' }}>
-          <LinkStyle
+          <Button
+            compact
+            noBackground
+            noBorder
+            highlightOnHoverAlt
+            padding="4px 12px"
             highlighted={highlightedIndex === 0}
             onClick={() => setHighlightedIndex(val => val === 0 ? null : 0)}
             onMouseEnter={() => setHighlightedIndex(val => val !== null ? 0 : null)}
@@ -374,7 +380,7 @@ function FileHeaderMenu({
             <Text>
               File
             </Text>
-          </LinkStyle>
+          </Button>
 
           <FlyoutMenu
             items={fileItems}
@@ -386,28 +392,12 @@ function FileHeaderMenu({
         </div>
 
         <div style={{ position: 'relative' }}>
-          <LinkStyle
-            highlighted={highlightedIndex === 1}
-            onClick={() => setHighlightedIndex(val => val === 1 ? null : 1)}
-            onMouseEnter={() => setHighlightedIndex(val => val !== null ? 1 : null)}
-            ref={refRun}
-          >
-            <Text>
-              Run
-            </Text>
-          </LinkStyle>
-
-          <FlyoutMenu
-            items={runItems}
-            onClickCallback={() => setHighlightedIndex(null)}
-            open={highlightedIndex === 1}
-            parentRef={refRun}
-            uuid="FileHeaderMenu/run_items"
-          />
-        </div>
-
-        <div style={{ position: 'relative' }}>
-          <LinkStyle
+          <Button
+            compact
+            noBackground
+            noBorder
+            highlightOnHoverAlt
+            padding="4px 12px"
             highlighted={highlightedIndex === 2}
             onClick={() => setHighlightedIndex(val => val === 2 ? null : 2)}
             onMouseEnter={() => setHighlightedIndex(val => val !== null ? 2 : null)}
@@ -416,7 +406,7 @@ function FileHeaderMenu({
             <Text>
               Edit
             </Text>
-          </LinkStyle>
+          </Button>
 
           <FlyoutMenu
             items={editItems}
@@ -427,11 +417,42 @@ function FileHeaderMenu({
           />
         </div>
 
+        <div style={{ position: 'relative' }}>
+          <Button
+            compact
+            noBackground
+            noBorder
+            highlightOnHoverAlt
+            padding="4px 12px"
+            highlighted={highlightedIndex === 1}
+            onClick={() => setHighlightedIndex(val => val === 1 ? null : 1)}
+            onMouseEnter={() => setHighlightedIndex(val => val !== null ? 1 : null)}
+            ref={refRun}
+          >
+            <Text>
+              Run
+            </Text>
+          </Button>
+
+          <FlyoutMenu
+            items={runItems}
+            onClickCallback={() => setHighlightedIndex(null)}
+            open={highlightedIndex === 1}
+            parentRef={refRun}
+            uuid="FileHeaderMenu/run_items"
+          />
+        </div>
+
         {PipelineTypeEnum.INTEGRATION !== pipeline?.type
           && featureEnabled?.(featureUUIDs.NOTEBOOK_BLOCK_OUTPUT_SPLIT_VIEW)
           && (
           <div style={{ position: 'relative' }}>
-            <LinkStyle
+            <Button
+              compact
+              noBackground
+              noBorder
+              highlightOnHoverAlt
+              padding="4px 12px"
               highlighted={highlightedIndex === 3}
               onClick={() => setHighlightedIndex(val => val === 3 ? null : 3)}
               onMouseEnter={() => setHighlightedIndex(val => val !== null ? 3 : null)}
@@ -440,7 +461,7 @@ function FileHeaderMenu({
               <Text>
                 View
               </Text>
-            </LinkStyle>
+            </Button>
 
             <FlyoutMenu
               items={viewItems}
@@ -454,7 +475,9 @@ function FileHeaderMenu({
 
         {featureEnabled?.(featureUUIDs.COMPUTE_MANAGEMENT) && (
           <div style={{ position: 'relative' }}>
-            <LinkStyle
+            <Button
+              compact
+              small
               highlighted={highlightedIndex === INDEX_COMPUTE}
               onClick={() => setHighlightedIndex(val => val === INDEX_COMPUTE ? null : INDEX_COMPUTE)}
               onMouseEnter={() => setHighlightedIndex(val => val !== null ? INDEX_COMPUTE : null)}
@@ -463,7 +486,7 @@ function FileHeaderMenu({
               <Text>
                 Compute
               </Text>
-            </LinkStyle>
+            </Button>
 
             <FlyoutMenu
               items={computeItems}
